Feat: 장바구니 선택 삭제 확인 모달 기능 구현 #25
장바구니 선택 아이템 개수에 따라 모달을 조건부 렌더링하는 방식으로 변경함.
이에 따라, 삭제 버튼 클릭 이벤트에서 아이템의 개수를 확인할 필요가 없기 때문에
모달을 오픈하는 함수 실행 코드만 추가함.
